Controller y Service de Booking
realizado el controller y service. Falta por probar, aun falta el controller y service del espacio comun y los middleware

+++ backend/src/services/booking.service.js
@@ -10,8 +10,8 @@ import { handleError } from "../utils/errorHandler.js";
 async function getAllBookings(req, res) {
     try {
         const bookings = await Booking.find().exec();
-        if(!bookings) return [null, "No se encontraron reservaciones"];
-        
+        if (!bookings) return [null, "No se encontraron reservaciones"];
+
     } catch (error) {
         handleError(error, "booking.service -> getAllBookings");
         return [null, error.message];
@@ -24,7 +24,7 @@ async function getAllBookings(req, res) {
 async function getBookingById(id) {
     try {
         const booking = await Booking.findById(id).exec();
-        if(!booking) return [null, "No se encontró la reservación"];
+        if (!booking) return [null, "No se encontró la reservación"];
 
         return [booking, null];
     } catch (error) {
@@ -39,30 +39,36 @@ async function getBookingById(id) {
 async function createBooking(req) {
     try {
         let email = req.body.email;
-        let bookingData = req.body;
+        const { spaceId, date, startTime, endTime } = req.body;
 
         // Verificar si el email ya está registrado
         let userId = await User
             .findOne({ email: email })
             .select("_id")
             .exec();
-        if(!userId) return [null, "El email no está registrado"];
+        if (!userId) return [null, "El email no está registrado"];
 
-        // agregar el id de usuario a la reservación
-        bookingData.userId = userId;
-
-        //Realizar validaciones de fechas de reservas
-        let bookingDate = new Date(bookingData.bookingDate);
-        let currentDate = new Date();
-        if(bookingDate < currentDate) return [null, "La fecha de reserva no puede ser anterior a la fecha actual"];
-
+        // Verificar si el espacio existe
+        const spaceExists = await CommonSpace.findById(spaceId);
+        if (!spaceExists) return [null, "El espacio no existe"];
         // Verificar si la fecha de reserva ya está ocupada
 
-        // Crear la reservación
-        const booking = new Booking(bookingData);
-        await booking.save();
+        const bookings = await Booking.find({
+            spaceId: spaceId,
+            date: date,
+            $or: [
+                { startTime: { $lt: endTime, $gte: startTime } },
+                { endTime: { $gt: startTime, $lte: endTime } }
+            ]
+        });
 
-        return [booking, null];
+        if (bookings.length > 0) return [null, "El espacio ya está reservado"];
+
+        // Crear la reserva
+        const newBooking = new Booking({ userId, spaceId, date, startTime, endTime });
+        await newBooking.save();
+
+        return [newBooking, null];
     } catch (error) {
         handleError(error, "booking.service -> createBooking");
         return [null, error.message];
@@ -72,10 +78,24 @@ async function createBooking(req) {
 /**
  * Actualiza una reservación existente.
  */
-async function updateBooking(id, bookingData) {
+async function updateBooking(id, req) {
     try {
+        let email = req.body.email;
+        // Verificar si el email ya está registrado
+        let userId = await User
+            .findOne({ email: email })
+            .select("_id")
+            .exec();
+        if (!userId) return [null, "El email no está registrado"];
+
+        //validacion el userId con el id de usuario de la reservacion
+        const bookingUser = await Booking.findById(id).select("userId").exec();
+        if (userId != bookingUser.userId || userId.roles != "admin") return [null, "El email no coincide con el usuario de la reservación"];
+
+        const { spaceId, date, startTime, endTime } = req.body;
+        const bookingData = { spaceId, date, startTime, endTime };
         const booking = await Booking.findByIdAndUpdate(id, bookingData, { new: true }).exec();
-        if(!booking) return [null, "No se encontró la reservación"];
+        if (!booking) return [null, "No se encontró la reservación"];
 
         return [booking, null];
     }
@@ -88,10 +108,22 @@ async function updateBooking(id, bookingData) {
 /**
  * Elimina una reservación.
  */
-async function deleteBooking(id) {
+async function deleteBooking(id, req) {
     try {
+        let email = req.body.email;
+        // Verificar si el email ya está registrado
+        let userId = await User
+            .findOne({ email: email })
+            .select("_id")
+            .exec();
+        if (!userId) return [null, "El email no está registrado"];
+
+        //validacion el userId con el id de usuario de la reservacion 
+        const bookingUser = await Booking.findById(id).select("userId").exec();
+        if (userId != bookingUser.userId || userId.roles != "admin") return [null, "El email no coincide con el usuario de la reservación"];
+
         const booking = await Booking.findByIdAndDelete(id).exec();
-        if(!booking) return [null, "No se encontró la reservación"];
+        if (!booking) return [null, "No se encontró la reservación"];
 
         return [booking, null];
     } catch (error) {
@@ -107,4 +139,3 @@ export default {
     updateBooking,
     deleteBooking
 };
-
